requirements.txt
python-telegram-bot==20.7
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0
python-dotenv==1.0.0
google-generativeai==0.3.2
pytz==2023.3.post1